backend/requirements.txt
# Core FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# OpenAI and Pinecone
openai==1.10.0
pinecone-client==3.0.2

# Redis Cache
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# Configuration & Environment
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Utilities
tenacity==8.2.3

# Async support
# asyncio is built into Python 3.4+, no need to install
aiofiles==23.2.1

# Testing & HTTP Client
httpx==0.26.0
pytest==7.4.3
pytest-asyncio==0.23.3

# Data Collection
playwright==1.41.0
beautifulsoup4==4.12.3
reportlab==4.0.9
